cleanup linux build process
