--- client_overlay/bin/x64/backup.desktop.txt
@@ -1,6 +1,0 @@
-[Desktop Entry]
-Type=Application
-Categories=Game;Utility;
-Exec=OpenVR-WalkInPlaceOverlay %F
-Icon=default
-Name=OpenVR-WalkInPlace
