+++ client_overlay/bin/x64/default.desktop
@@ -1,6 +1,6 @@
 [Desktop Entry]
 Type=Application
-Categories=Game;Utility;
-Exec=OpenVR-WalkInPlaceOverlay %F
-Icon=ovrwip
+Categories=Game;
+Exec=AppRun
+Icon=default
 Name=OpenVR-WalkInPlace
