+++ client_overlay/bin/x64/res/qml/MyToggleButton.qml
@@ -7,24 +7,24 @@ CheckBox {
 	hoverEnabled: true
 	spacing: 12
 
-	indicator: Rectangle {
-		implicitWidth: 28
-		implicitHeight: 28
-		x: parent.leftPadding
-		y: parent.height / 2 - height / 2
-		color: parent.activeFocus ? myPalette.mid : myPalette.base
-		border.width: 0
-		Image {
-			width: 38
-			height: 38
-			x: (parent.width - width) / 2
-			y: (parent.height - height) / 2
-			source: "image://default/check/#ffffff"
-			sourceSize.width: width
-			sourceSize.height: height
-			visible: parent.parent.checked
-		}
-	}
+	// indicator: Rectangle {
+	// 	implicitWidth: 28
+	// 	implicitHeight: 28
+	// 	x: parent.leftPadding
+	// 	y: parent.height / 2 - height / 2
+	// 	color: parent.activeFocus ? myPalette.mid : myPalette.base
+	// 	border.width: 0
+	// 	Image {
+	// 		width: 38
+	// 		height: 38
+	// 		x: (parent.width - width) / 2
+	// 		y: (parent.height - height) / 2
+	// 		source: "image://default/check/#ffffff"
+	// 		sourceSize.width: width
+	// 		sourceSize.height: height
+	// 		visible: parent.parent.checked
+	// 	}
+	// }
 
 	contentItem: MyText {
         text: parent.text
