--- driver_vrwalkinplace/src/dllmain.cpp
@@ -1,46 +1,0 @@
-// dllmain.cpp : Definiert den Einstiegspunkt für die DLL-Anwendung.
-#include "logging.h"
-
-const char* logConfigFileName = "logging.conf";
-
-const char* logConfigDefault =
-"* GLOBAL:\n"
-"	FORMAT = \"[%level] %datetime{%Y-%M-%d %H:%m:%s}: %msg\"\n"
-"	FILENAME = \"driver_vrwalkinplace.log\"\n"
-"	ENABLED = true\n"
-"	TO_FILE = true\n"
-"	TO_STANDARD_OUTPUT = true\n"
-"	MAX_LOG_FILE_SIZE = 2097152 ## 2MB\n"
-"* TRACE:\n"
-"	ENABLED = false\n"
-"* DEBUG:\n"
-"	ENABLED = false\n";
-
-INITIALIZE_EASYLOGGINGPP
-
-void init_logging() {
-	el::Loggers::addFlag(el::LoggingFlag::DisableApplicationAbortOnFatalLog);
-	el::Configurations conf(logConfigFileName);
-	conf.parseFromText(logConfigDefault);
-	conf.parseFromFile(logConfigFileName);
-	conf.setRemainingToDefault();
-	el::Loggers::reconfigureAllLoggers(conf);
-}
-
-BOOL APIENTRY DllMain( HMODULE hModule,
-                       DWORD  ul_reason_for_call,
-                       LPVOID lpReserved
-					 ) {
-	switch (ul_reason_for_call) {
-		case DLL_PROCESS_ATTACH:
-			init_logging();
-			LOG(INFO) << "VRWalkInPlace dll loaded...";
-			break;
-		case DLL_THREAD_ATTACH:
-		case DLL_THREAD_DETACH:
-		case DLL_PROCESS_DETACH:
-			break;
-	}
-	return TRUE;
-}
-
