+++ driver_vrwalkinplace/src/driver/ServerDriver.cpp
@@ -3,6 +3,32 @@
 
 #include <boost/date_time/posix_time/posix_time_types.hpp>
 
+constexpr char logConfigFileName[] = "logging.conf";
+
+constexpr char logConfigDefault[] =
+"* GLOBAL:\n"
+"	FORMAT = \"[%level] %datetime{%Y-%M-%d %H:%m:%s}: %msg\"\n"
+"	FILENAME = \"driver_vrwalkinplace.log\"\n"
+"	ENABLED = true\n"
+"	TO_FILE = true\n"
+"	TO_STANDARD_OUTPUT = true\n"
+"	MAX_LOG_FILE_SIZE = 2097152 ## 2MB\n"
+"* TRACE:\n"
+"	ENABLED = true\n"
+"* DEBUG:\n"
+"	ENABLED = true\n";
+
+INITIALIZE_EASYLOGGINGPP
+
+void init_logging() {
+	el::Loggers::addFlag(el::LoggingFlag::DisableApplicationAbortOnFatalLog);
+	el::Configurations conf(logConfigFileName);
+	conf.parseFromText(logConfigDefault);
+	conf.parseFromFile(logConfigFileName);
+	conf.setRemainingToDefault();
+	el::Loggers::reconfigureAllLoggers(conf);
+}
+
 namespace vrwalkinplace {
 	namespace driver {
 
@@ -21,6 +47,7 @@ namespace vrwalkinplace {
 		}
 
 		vr::EVRInitError ServerDriver::Init(vr::IVRDriverContext *pDriverContext) {
+			init_logging();
 			LOG(TRACE) << "CServerDriver::Init()";
 
 			LOG(DEBUG) << "Initialize driver context.";
