+++ run.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-cd ~/.local/share/vrwalkinplace/
+cd ~/.steam/steam/vrwalkinplace/
 
-~/.steam/steam/ubuntu12_32/steam-runtime/run.sh ./OpenVR-WalkInPlaceOverlay -nomanifest
+./OpenVR-WalkInPlaceOverlay
